verzia 17 +users

+++ _inc/delete-item.php
@@ -2,7 +2,10 @@
 
 	// include
 	require 'config.php';
-
+	if ( ! logged_in() ) {
+		header("Location: $base_url/login.php");
+		die();
+	}
 
 	// delete item
 	$affected = $database->delete('mylinks',
